+++ DX12App/UI.cpp
@@ -77,7 +77,7 @@ void UI::BeginDraw(UINT nFrame)
 void UI::TextDraw(UINT nFrame, UINT TextCnt, const std::vector<TextBlock> &mvTextBlocks)
 {
     //0번은 테두리 색.
-    for (int i =0; i < TextCnt; ++i)
+    for (int i =0; i < static_cast<int>(TextCnt); ++i)
     {
         mpd2dDeviceContext.Get()->DrawTextW(mvTextBlocks[i].strText.c_str(), static_cast<UINT>(mvTextBlocks[i].strText.length()),
             mvdwTextFormat[i].Get(), mvTextBlocks[i].d2dLayoutRect, mvd2dSolidBrush[i+1].Get());
@@ -88,7 +88,7 @@ void UI::RectDraw(XMFLOAT4 RectLTRB[], XMFLOAT4 FillLTRB[], UINT TextCnt, UINT n
 {
     if (md2dLinearGradientBrush.Get())
     {
-        for (int i = 0; i < GradientCnt; ++i)
+        for (int i = 0; i < static_cast<int>(GradientCnt); ++i)
         {
             mpd2dDeviceContext.Get()->FillRectangle(D2D1::RectF(FillLTRB[i].x, FillLTRB[i].y, FillLTRB[i].z, FillLTRB[i].w), md2dLinearGradientBrush.Get());
             mpd2dDeviceContext.Get()->DrawRectangle(D2D1::RectF(RectLTRB[i].x, RectLTRB[i].y, RectLTRB[i].z, RectLTRB[i].w), mvd2dSolidBrush[0].Get());
@@ -106,7 +106,7 @@ void UI::RoundedRectDraw(XMFLOAT4 RectLTRB[], XMFLOAT4 FillLTRB[], UINT TextCnt,
 {
     if (md2dLinearGradientBrush.Get())
     {
-        for (int i = 0; i < GradientCnt; ++i)
+        for (int i = 0; i < static_cast<int>(GradientCnt); ++i)
         {
             mpd2dDeviceContext.Get()->FillRoundedRectangle(D2D1::RoundedRect(D2D1::RectF(FillLTRB[i].x, FillLTRB[i].y, FillLTRB[i].z, FillLTRB[i].w), 10.0f, 10.0f), md2dLinearGradientBrush.Get());
             mpd2dDeviceContext.Get()->DrawRoundedRectangle(D2D1::RoundedRect(D2D1::RectF(RectLTRB[i].x, RectLTRB[i].y, RectLTRB[i].z, RectLTRB[i].w), 10.0f, 10.0f), mvd2dSolidBrush[0].Get());
@@ -165,9 +165,9 @@ void UI::CreateFontFormat(float FontSize, const std::vector<std::wstring> &Fonts
 {
     mvdwTextFormat.resize(TextCnt);
 
-    for (int i = 0; i < TextCnt; ++i)
+    for (int i = 0; i < static_cast<int>(TextCnt); ++i)
         ThrowIfFailed(mpd2dWriteFactory->CreateTextFormat(Fonts[i].c_str(), nullptr, DWRITE_FONT_WEIGHT_NORMAL, DWRITE_FONT_STYLE_NORMAL, DWRITE_FONT_STRETCH_NORMAL, FontSize, L"en-us", mvdwTextFormat[i].GetAddressOf()));
-    for (int i = 0; i < TextCnt; ++i)
+    for (int i = 0; i < static_cast<int>(TextCnt); ++i)
     {
         ThrowIfFailed(mvdwTextFormat[i]->SetTextAlignment(Alignment[i]));
         ThrowIfFailed(mvdwTextFormat[i]->SetParagraphAlignment(DWRITE_PARAGRAPH_ALIGNMENT_CENTER)); // DWRITE_PARAGRAPH_ALIGNMENT_NEAR
@@ -186,7 +186,7 @@ void UI::BuildSolidBrush(UINT UICnt, UINT TextCnt, D2D1::ColorF* ColorList)
     //첫번째 SolidColorBrush는 무조건 Black으로 설정한 후 테두리로 사용할 것. 그게 편할 듯.
     mvd2dSolidBrush.resize(TextCnt+UICnt);
     ThrowIfFailed(mpd2dDeviceContext->CreateSolidColorBrush(D2D1::ColorF(D2D1::ColorF::Black), (ID2D1SolidColorBrush**)&mvd2dSolidBrush[0]));
-    for (int i = 0; i < TextCnt+UICnt-1; ++i)
+    for (int i = 0; i < static_cast<int>(TextCnt+UICnt-1); ++i)
         ThrowIfFailed(mpd2dDeviceContext->CreateSolidColorBrush(D2D1::ColorF(ColorList[i]), (ID2D1SolidColorBrush**)&mvd2dSolidBrush[i+1]));
     // 0번 SolidBrush는 무조건 Black, 나머지는 인자로 받은 ColorList로 설정. 따라서 Resize할 때 UI와 Text 수에다가 1을 더해서 설정
     // 1번부터는 ColorList색. Text색 이후 UI 색으로 설정
@@ -196,7 +196,7 @@ void UI::BuildSolidBrush(UINT UICnt, UINT TextCnt, std::vector<D2D1::ColorF>& Co
 {
     mvd2dSolidBrush.resize(TextCnt + UICnt);
     ThrowIfFailed(mpd2dDeviceContext->CreateSolidColorBrush(D2D1::ColorF(D2D1::ColorF::Black), (ID2D1SolidColorBrush**)&mvd2dSolidBrush[0]));
-    for (int i = 0; i < TextCnt + UICnt - 1; ++i)
+    for (int i = 0; i < static_cast<int>(TextCnt + UICnt - 1); ++i)
         ThrowIfFailed(mpd2dDeviceContext->CreateSolidColorBrush(D2D1::ColorF(ColorList[i]), (ID2D1SolidColorBrush**)&mvd2dSolidBrush[i + 1]));
 }
 
@@ -205,7 +205,7 @@ void UI::BuildLinearGradientBrush(UINT ColorCnt, D2D1::ColorF* ColorList)
     ID2D1GradientStopCollection* pGradientStops = NULL;
     D2D1_GRADIENT_STOP* gradientStops = new D2D1_GRADIENT_STOP[ColorCnt];
     
-    for (int i = 0; i < ColorCnt; ++i)
+    for (int i = 0; i < static_cast<int>(ColorCnt); ++i)
     {
         gradientStops[i].color = ColorList[i];
         gradientStops[i].position = static_cast<float>(i) * 1.0f / 3.0f;
